Pushing new formulation with tg plane at edge for spherical case
Also added a times 2 on the metric term for stressDivergenceUCGrid
otherwise convergence was not reached wrt the analytic soln.
Now, with the alternate denominator, convergence is 2nd order in
 the L2 norm and first order in the Linfinity.
Need to find the right correction for the metric terms in stressDivergenceVCGrid.
Note that there was no convergence with the original denominator