Merge branch 'dylanschlichting/ocn/add-surface-divergence' (PR #6892)
Add ocean divergenceAtSurface to highFrequencyStats

This commit mirrors PR #6566, but adds divergence at the surface. This
is useful for MPAS-O meshes capable of permitting or resolving
submesoscale processes. Divergence is a key diagnostic of submesoscale
flows, in particular identifying regions of frontogenesis and
frontolysis. The variable is added to the default output for
highFrequencyStats, but is an optional 2D field.

[BFB]